[TASK] Extract SVG image processing into processor
Instead of spreading (similar) code that handles SVG
files in multiple classes with conditional code
branching, processing SVG files (mostly calculating
new width and height according to the configuration)
is now placed in an own file processor.

Cropping SVGs is still done like before using IM/GM
and resulting in an png image. In this case the new
processor is skipped and the file is passed to the
existing LocalImageProcessor.

A new class is introduced, which calculates resulting
image dimension from an original dimension and a given
processing task.

Releases: master, 10.4
Resolves: #92014
Change-Id: I5799c09b06d32a495a917da523506a6cc1f8e45d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65556
Tested-by: Helmut Hummel <typo3@helhum.io>
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>