[BUGFIX] Register deprecated icon if target extension is loaded
The IconRegistry contains two deprecated icons whose replacements
are located in ext:impexp, which is not a hard required extension.
Calling IconRegistry->getAllRegisteredIconIdentifiers(), iterating
over those icons - for example to render them - fails with an
exception "Icon with identifier ... is not registered" if ext:impexp
is not loaded.
The deprecations have been removed in master, so only v10 is affected.
The patch adds those two deprecations only if ext:impexp is loaded.

Resolves: #92021
Releases: 10.4
Change-Id: I185f0b1a5a75ef643474a1794af7cab436182839
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65343
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Haupt <mail@danielhaupt.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Daniel Haupt <mail@danielhaupt.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>