[BUGFIX] Drop 'List View' in workspace preview
The workspace preview not only has the default view with
the slider to show differences between live and workspace
version. It additionally contains a full workspace module
accessible through the 'List View' button.
This view is notoriously broken and not fully thought through.
For instance the 'Open version of page' button breaks out of
the module to a different place of the backend, restricted to
the main module frame, and there is no way back.
For simple workspace stage editing, the default preview
mode has the 'send to stage' and 'discards' buttons already,
so there is little reason to have the backend module rendered
in the preview in the first place.
The patch drops the 'List View' from the preview. Next to
a cleaner interface, this increases the performance of the
preview since a full backend call plus various ajax calls
are no longer executed.

Resolves: #92031
Releases: master, 10.4
Change-Id: I33d21931e157aed2ef64cca3808cc4b6c28a8ef1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65352
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>