[BUGFIX] Keep FE cookie expire time correct
Using FE permalogin sets a cookie expire time once. While the session
timestamp on the server side is updated continuously to keep a session
alive. This led to the situation that an actually valid session was
abandoned due to cookie expiry.

The code is adjusted to always update an existing cookie if it is a
non-session cookie.

Resolves: #61074
Releases: master, 10.4
Change-Id: I3a0eb9888546e3906a572540b69795ae5ad2365e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65356
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>