[TASK] Replace string parsing in Message with unpack
Replaces an inefficient string parsing using strlen, for
and ord, to use unpack instead. Saves one call to strlen
per header, and potentially thousands of calls to ord
per request.

The same strategy is used in Message->filter, however,
this method does not appear to be used so the unpack
change is only applied in isValidHeaderValue.

Releases: master, 10.4
Resolves: #92059
Change-Id: I6f485287eb6773accd8bd194ec998205479cec5f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65396
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>