[TASK] DatabaseRecordList: Drop unnecessary DB-Query for each table
For each TCA table a query to determine if a result exists is made.
It could have been some kind of performance optimization,
but it results in an additional query for each table with records
in place.

Such optimizations should be done in the getTable method by using the
existing query to count the total number of records of the table.
Therefore, the query result (the total number of records of the table)
has been used for a guard clause.

Releases: master
Resolves: #92071
Change-Id: Idc2026eaaa699ab358625d00670a5bc1e8ab341f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65410
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>