[BUGFIX] DB check -> full search shows deleted records if requested
With #82729 the $GLOBALS['SOBE'] = $this; assignment in
DatabaseIntegrityController has been removed. This is a
good thing, however parts of QueryGenerator class still
relied on this.
The patch gives the module settings that have been
introduced for #83428 from QueryView down to
QueryGenerator as well.
This fixes a couple of details, most importantly the
'Show even deleted entries' checkbox of the full search
advanced query view works again.

Resolves: #92079
Related: #83428
Related: #82729
Releases: master, 10.4
Change-Id: If8b5ef9e2d86910625455e37d029c5e93e5ef519
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65480
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>