[TASK] Merge QueryView with QueryGenerator and move to lowlevel
The classes QueryView and QueryGenerator in the core extension
are detail implementations of the databases search functionality
of the lowlevel extension.
The patch merges both classes into one, moves it to the lowlevel
extension and deprecates the old classes. Minor refactorings like
protecting properties and methods is applied to the new class to
enable further refactoring if desired.

Change-Id: If8b16e8eeb39ed1e3d3aa52eed02b6c27a114e1a
Resolves: #92129
Related: #92080
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65457
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>