[BUGFIX] Import data from yaml imports in declaration order
For plain arrays it's sometimes important that the items
appear in the order in which they are imported. For example
with the languages in the sites configuration where the first
language is seen as the default language.

Just like with Symfony Service Container imports come first
and its settings are overwritten with the main file. To make
sure that multiple imports are loaded in the order as they
appear in the main file they must be processed in reverse
order.

This behaviour can now be controlled by a new feature toggle
which defaults to `false` for existing and to `true` for new
installations.

Resolves: #92100
Releases: master, 10.4
Change-Id: I30ffba171606f3c62c15742eba3187af454aa5b6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68074
Tested-by: core-ci <typo3@b13.com>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>