[!!!][TASK] DatabaseRecordList: Drop hook to modify `searchFields`
The TCA configuration `searchFields` in the `ctrl` section has been
introduced in TYPO3 4.6.
This configuration allows setting the search columns. Those columns are
taken into account by the search in the TYPO3 backend.

To enable a smooth transition between TYPO3 4.5 and 4.6, the hook
`$GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['mod_list']
['getSearchFieldList']` has been introduced as well. It allows
to set the `searchFields` for the list module's search.

As this transition should be finished now, the hook has been
removed. Extension authors are encouraged to set the `searchFields` via
the TCA configuration.

Releases: master
Resolves: #92128
Change-Id: I3afef4f248797345e4feac14a10a7c7797bc86e3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65502
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>