[TASK] Streamline usage of contrib packages in TypeScript
No need to use a fully qualified namespace for
3rd party modules that are placed in TYPO3/CMS/Core/Contrib.
There are usually aliases configured in the requirejs configuration,
and if they are missing, we add them now.

This change additionally drops an unneded module declaration
for "TYPO3/CMS/Core/Contrib/imagesloaded.pkgd.min" which is
not needed when simply using "imagesloaded" as module name.

Commands used:

  grunt build

Releases: master, 10.4
Resolves: #92725
Change-Id: I6e7c02104050202d5c1a8bd0d7546a1496f5636c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65693
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>