[!!!][FEATURE] Improve ValidatorTask
The ext:linkvalidator ValidatorTask has been improved. Therefore
the task was switched from marker based templates to FluidEmail.

The mail content was extended for a proper list of broken links
along with all necessary information such as record id, page id,
and so on. This allows editors now to easily access the records
with broken links without having to visit the backend module first.

Because of heavy refactoring and introduction of a new PSR-14
event, extension authors are now able to fully customize the mail
content and the configuration (such as to, from, subject) dynamically
(e.g. adding another recipient only on production context).

Both the task and its field provider got added strict types. The
task however currently still needs some type casts for backwards
compatibilty as the tasks state is stored as serialized string in
the database.

The whole link analyzing process was moved into the dedicated class
`LinkAnalyzerResult`.

A new field `languages` is introduced which allows to restrict
the report to the defined languages which is handy if different
tasks should be used (e.g. editors responsible for only one
localization).

Resolves: #29342
Releases: master
Change-Id: Ibe070038ef66f876de38e47d6f82f4f090af96ce
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65549
Tested-by: Sybille Peters <sypets@gmx.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Sybille Peters <sypets@gmx.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>