[TASK] DatabaseRecordList: Streamline `returnUrl`
The method `listURL()` is used to set the `returnUrl` for all actions
in the `DatabaseRecordList`.
It replaces the usages of the method `makeReturnUrl`. Therefore, this
method has been removed. Furthermore,
`GeneralUtility::getIndpEnv('REQUEST_URI')` has been replaced by
`listURL()`.
This leads to a consistent behaviour after performing any kind of
action in the `DatabaseRecordList`.

Releases: master
Resolves: #92192
Resolves: #83758
Change-Id: I2dc84a2c4e4103d5b4e77ebb646158e8c15f3202
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65574
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>