[BUGFIX] Don't call ResourceFactory::getStorageObject twice
In http://review.typo3.org/11011 a plain DB call
that fetched the storage record was replaced with
usage of StorageRepository::findByUid, which in
turn calls ResourceFactory::getStorageObject again.

This leads to all events also called twice.

To avoid this a new internal method is introduced
on StorageRepository, which only fetches the storage
row again.

Resolves: #92201
Releases: master, 10.4, 9.5
Change-Id: Ibce6e8d9a0b5ce7abb0f86d80ac42a1a86d47e35
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65594
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Helmut Hummel <typo3@helhum.io>