[TASK] Provide web server url configuration for acceptance tests
In order to run TYPO3 Core acceptance tests not only in the TYPO3 Core
own docker environment but in an arbitrary environment the web server
url of the testing environment should be configurable by the
environment variable "typo3TestingAcceptanceBaseUrl".

For example, having a LEMP stack with an Nginx web server available at
http://nginx:8080 you would set the variable
"typo3TestingAcceptanceBaseUrl=http://nginx:8080" in the environment
of the acceptance tests runner.

Resolves: #92205
Releases: master, 10.4
Change-Id: Ia890ed813d1e46af0b9010f325b13f08f9863ab2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65585
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>