[BUGFIX] PlainDataResolver misses DeletedRestriction
The PlainDataResolver is used to find overlay id's of live
id's in workspace context. One usage is the ReferenceIndex.

In scenarios 'modify content, discard content, modify again'
with relations, the PlainDataResolver sometimes fetches rows
that have been set to deleted (discarded) already together
with the rows that should be fetched. Depending on the order
of the result rows - which depends on the DBMS - wrong overlay
rows are calculated sometimes.

The patch adds a database DeletedRestriction at 3 of 4 places
in the PlainDataResolver to only fetch correct rows. The last
place triggers another side-effect that needs to be tackled
with another patch and is commented as @todo for now.

The tests add missing sys_refindex rows to the 6 tests that
failed on postgres before. Additionally, for some other
IRRE/CSV workspace tests, the sys_refindex changes slightly
and is more correct now.

Resolves: #92190
Releases: master, 10.4
Change-Id: I14f5211b5996d7f53738b9e086c3866aa1a4eaeb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65617
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>