[TASK] Drop workspace publish locking
The workspace 'publish' process is encapsulated in a lock
file to prevent dupblicate publish of the same record at
the same time.

This is bogus:
* Other operations like modify or discard are not covered
* A stalled lock file can only happen if a process crashes
* Nothing else in the DataHandler has similar locking strategies
* The only way to recover from a stalled lock file is "ask your admin",
  which is not helpful in any way.

If the DataHandler (and with that the workspace operations) should
have locking, it needs to be implemented in the DataHandler at a
central place and not within a fraction of the workspace sub set.

The patch drops the publish process locking.

Change-Id: I4004f46eb3dc47bb62347f5bf45a9ed0c6f5da1e
Resolves: #92240
Releases: master, 10.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65647
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>