[BUGFIX] Use correct variable to access language id of a column
The fluid based page module used a non existing
variable to access the language id of a column. This lead
to passing `NaN` as language id to the DataHandler when
moving a content element in the corresponding column.
Therefore the content element was actually moved to the
default column because of the changed language.

The template is now accessing the correct variable fetched from
the SiteLanguage object which makes moving content on translated
pages possible again.

Resolves: #92086
Releases: master, 10.4
Change-Id: I6fb95338e47e9ad5b0a3f637f6b637174271ba35
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65691
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Riccardo De Contardi <erredeco@gmail.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>