[BUGFIX] Do not show editlock fields to editors
Editors should never see the "editlock" field, which locks
the record for editing of admins only.

In order to achieve this, the fields pages.editlock and
tt_content.editlock are changed to never be shown for non-admins
and on top, the field is not shown in setting permissions
within be_groups.allowed_exclude_fields anymore.

Resolves: #87100
Releases: master, 10.4
Change-Id: I39b9212f0e955fd1a284b15014fcf00078cab16e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65786
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>