[TASK] Add functional tests for RootlineUtility
This change introduces functional tests for the rootline utility.

For now, this patch ensures the validity of Workspace
Overlays when reading a given page tree / rootline.

Covered tests:
* Live workspace (no overlays)
* New record in workspace
* Modified record in workspace
* Deleted record in workspace
* Moved record in workspace

Adding such a test class helps to extend
coverage for language overlays and MountPoints
in conjunction with workspaces in the future.

Resolves: #92311
Releases: master, 10.4
Change-Id: Ie8a746253e4b10f98c05a448d736a5d00d59d3fc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65741
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>