[TASK] Enable EXT:impexp to include presets in export
Only tables with TCA configurations are available for the export of a
TYPO3 instance. The stored export configurations - so called presets -
were missing the TCA configuration and thus had to be exported and
imported separately in a database client.

This patch adds a TCA configuration for the export configurations and
thus makes them exportable and importable along with the main dump.
The presets are saved on PID=0 and hidden from record lists to
force the user to continue managing them with the export module.

Resolves: #92346
Releases: master, 10.4
Change-Id: Ic5a9babc91a93f8bf1561b697c4fca0ad548f734
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65770
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Rémy DANIEL <dogawaf@no-log.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>