[BUGFIX] Skip sys_refindex for workspace placeholders
The reference index can be seen as a cache for relations.
It is used in various places to show how many other records
have a relation to a given record, for instance in list
and filelist module and in the record info modal.

In workspaces, new and moved records create records pairs:
t3ver_state -1 & 1 and t3ver_state 3 & 4. The backend
always shows only one of these pairs. The reference index
creates an entry for each record, so two for each pair.
This is useless and confusing for editors: If for instance
a new file reference is added to a content element in
workspaces, the file list shows two usages of the file
instead of only one.

The patch suppresses reference index entries for the
placeholder records - t3ver_state 1 and 3 - so each pair
ends up with only one sys_refindex entry.

Resolves: #92345
Related: #61917
Releases: master, 10.4
Change-Id: I16ede3a9f1b66a7195526a224e9f1c43c03d7ba6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65767
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>