[BUGFIX] Correct sys_refindex update on workspace publish
When publishing workspace changes to live, many
sys_refindex rows are not updated properly, especially
if dealing with workspace-new or workspace-moved rows
and relations.

To update the reference index for relations, the data handler
uses class ReferenceIndex, which uses class RelationHandler,
which uses PlainDataResolver. All of these classes are
workspace aware and pick the current user workspace if
not explicitely told otherwise.

When publishing records, a given user is currently in some
workspace, but the reference index needs to be fixed for
live. This explicit runtime workspace switch is lost in
the ReferenceIndex.

The patch hands over given workspace context from
ReferenceIndex to RelationHandler during publish.

This one-liner fixes tons of broken reference index cases.
It also reveals a couple of still missing entries. Those
will be tackled with another patch.

Change-Id: Ib7647baad827fa8f148763efe91b331e5ad4484e
Resolves: #92348
Releases: master, 10.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65771
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>