[BUGFIX] Load dependencies of ImmediateActionElement dynamically
The module ImmediateActionElement is used to render a defined custom
HTML tag triggering certain actions once it's rendered, like reloading
the module menu or the backend's topbar. To achieve this, it needs
functionality from the modules TYPO3.ModuleMenu.App.refreshMenu and
TYPO3.Backend.Topbar.refresh.

This module is loaded at a very early stage where those dependencies
cannot do anything meaningful as there is nothing to do, namely the
backend login, which causes JavaScript warnings and increases loading
time.

To relax the situation, the modules are now loaded dynamically when
required by using async functions.

In a perfect world, there would be some kind of a hooking system and no
hardcoded list of actions, but we're not there, yet.

Resolves: #92350
Releases: master, 10.4
Change-Id: I50029789f564ce339b3204d51f61e4cc39bb3cf9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67312
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>