[BUGFIX] Update localized records when publishing default translation
If an editor only published the default language,
then the localized records are left "hanging in" a workspace,
and are marked as deleted, this usually happened
when moving records.

The use-case however, more often is to publish a default
language first and then update the translations
when they are ready.

This change updated the l10n_parent and l10n_source fields
for the records on publish to use the live ID then,
but can be kept in the workspace.

Resolves: #92445
Releases: master, 10.4
Change-Id: I5504833c5f7b4f30a3c79421c6f783295871f8c1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65979
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>