[BUGFIX] Restore recordlist localization behaviour
With #92133 the translation and localization behaviour within
recordlist was streamlined. Thereby some changes were made
that led to unexpected changes. These are now fixed.

The LocalizationPanel is not shown anymore for records which are
already localized in all possible languages, are in free mode or
have sys_language_uid -1 set.

This is achieved by taking the TranslationConfigurationProvider::translationInfo()
result into account again.

The indentation of the "Localization" column on localized rows
is added again. Furthermore, a misbehaviour is fixed in which
the rows' l18n_parent field was checked directly instead of the
defined transOrigPointerField field. As a result, the indentation
was only applied to tt_content rows as this is the only core table
having l18n_parent as transOrigPointerField configured.

Resolves: #92447
Relates: #92133
Releases: master
Change-Id: Ib9cc5af0bddc40d1546224e4b886135bda125a3f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65961
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Windloff
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Daniel Windloff
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>