[TASK] Rephrase explanation of sprintf arguments
This patch changes the conversion specification to `%s` while
keeping the reference to the PHP `sprintf()` function and
adds an example of how such a locallang label could look like.

Additionally, the documentation no longer gives the impression that
you can use `%1` (literally) in a language source file.

Releases: master, 10.4
Resolves: #92450
Change-Id: I37fc826b54c718b68625291b9c8544ac163c6e71
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65962
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Björn Jacob <bjoern.jacob@tritum.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Björn Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>