[BUGFIX] Catch exception during unserializing task
If the class can't be unserialized, an exception should
be caught to avoid having a state which is not solvable
via backend anymore.

The v12/main solution of this is different since we
rewrote the main controller in v12 meanwhile. This
backport is a much more straight solution that should
be suitable for v11.

Resolves: #92466
Resolves: #91766
Releases: main, 11.5
Change-Id: I2aa9d793c99b47b3d81599def034cea9af961ba6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73302
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>