[BUGFIX] Use correct page id for permission calculation of inline controls
For the inline record header controls the InlineRecordContainer calculates
the permissions for the page, the record is on. Previously the calculation
was based on the records pid. This does not work for new records as the
pid is a negative integer in this case - uid of the record the `new` button
was pressed on - and therefore led to no header controls were displayed
till the record was saved.

The calculation now uses the correct uid from `parentPageRow`.

Resolves: #84867
Releases: master, 10.4
Change-Id: I3d60f41ed51ff2f013d37c6aa9c118974d06acad
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65999
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>