[BUGFIX] Provide colPos for each cell in defLangBinding view
When dragging content elements in page module, the colPos
value of the dragged element is updated based on the dropped
area. A drop area is a table cell `td` in which the elements are
placed as `div`. The cell itself has the colPos set via a data
attribute. This means moving an element (`div`) to another
cell (`td`), the element gets updated with the colPos of the
new cell.

This did not work if `defLangBinding` was activated because
all elements are then assigned to separate cells which missed
the colPos information completely.

The colPos information is now added to all cells in the
`defLangBinding` view.

Resolves: #92454
Releases: master, 10.4, 9.5
Change-Id: I1ab5cb0b3dd5574b6d3397dc52397365dc59d1ee
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66015
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Benni Mack <benni@typo3.org>