[BUGFIX] Provide original message and envelope to transport
The interface Symfony\Component\Mailer\Transport\TransportInterface
expects a RawMessage and an optional Envelope if it's not already part
of RawMessage. Currently an instance of Symfony\Component\Mailer\SentMessage
is sent which is not compatible.

This patch changes this behavior to send the original message and the
envelope like already done in the MemorySpool.

Resolves: #92053
Releases: master, 10.4
Change-Id: I2eb1a6a40c192ac433c6cc112b05432018411312
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66058
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Markus Gerdes <markus@madaxel.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Markus Gerdes <markus@madaxel.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>