[TASK] Explicitly require PHPUnit 8
The code in the test suite is based on PHPUnit 8
therefore using it with PHPUnit 9 will produce a lot of warnings
caused by deprecations.

The testing framework however allows installing PHPUnit 9,
which is fine, when the test suite itself can cope with that.

This is why we pin our tests to PHPUnit 8, until we adapt
them to properly work with PHPUnit 9

composer req --dev phpunit/phpunit ^8.5.2 --no-update
composer update phpunit/phpunit

Resolves: #92511
Releases: master, 10.4
Change-Id: I9270dd6d7e8c0b8b6a8794a54abc2195c8aad896
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66077
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66082
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Helmut Hummel <typo3@helhum.io>