[TASK] Stop JsonView from setting headers
JsonView did set content type headers in its render method which
is a problem since the view should only be responsible for rendering
the content of the to be returned response.

The current approach, setting those headers in the action controller,
is just a temporary solution until Extbase is capable of handling
PSR-7 responses which then makes it superfluous to set headers
magically. Instead people will need to set headers manually or use
a JsonResponse object instead and let the core handle setting headers
further down in the request/response handling stack.

Releases: master
Resolves: #92692
Change-Id: Idfa417cfd25d7f692dea2e6d6822f1d3a44cb38d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66078
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>